chore: add variant suffix instead of prerelease version
